stock-ws module
Implementation of setStock and getLastStockId methods in
StockController.
Created WsProcessor service layer which will handle the controller's
requests.
WsProcessor makes use of the integration layer to map and post Stock
entity to be persisted. It also makes use of the data processor to
retrieve database information.
Updated integration-config.xml file with transformation and pollable
channels to be used to persist data in the database.