new changes including unit tests
Removed transformation channels since all it was doing was simple
mapping of objects which is already simple enough by using orika.
Added queryDSL functionality in order not to have any hardcoded sql.
Changed getLastStock to return the whole Stock entity instead of only
returning its relevant id.
Added a new WS object, ShowStock which is and extension of the Stock WS
object with the added id value which is being returned by the entity
once persisted.
Added unit tests to verify system integrity and robustness.
Test cases include validating input parameters and ascertaining that the
entities are being persisted in the database.