Added gateway , caching, Stock return on persist.
Added new Gateway interface and implementation which handles the
integration channel.
Thanks to the gateway layer, two way communication is now possible and
the persisted Stock entity object can be returned so that consumer can
know the id of the persisted object.
Added caching bean in order to evade unnecessary querying for the last
stock provided that no new stock has been persisted.
Removed Integrator class which was replaced by the Gateway.