Updated StockControllerValidation, WsProcessor
StockControllerValidation is now intercepting errors of type Exception
since the web service was unable to read such messages which could occur
when database is unavailable.
WsProcessor was updated to handle null responses from the gateway. These
are caught and raised as a timeout.
WsProcessorTest test class to test timeout scenario with gateway.